Refactor installer into own module
